Decompose stock fetching logic into hook
